refactor: extract SortableCostItem and fix TypeScript errors
- Extracted SortableCostItem component to its own file for better maintainability
- Fixed TypeScript errors related to event handling and type imports
- Updated imports and type definitions for better type safety
- Improved UI by replacing drag handle with GripVertical icon
