+++ components/sortable-cost-item.tsx
@@ -1,0 +1,273 @@
+"use client";
+
+import React from "react";
+import { formatNumber } from "@/utils/format";
+import { createPortal } from "react-dom";
+import { useSortable } from '@dnd-kit/sortable';
+import { CSS } from '@dnd-kit/utilities';
+import { PlusCircle, Trash2, GripVertical } from "lucide-react";
+import { Button } from "./ui/button";
+import { Input } from "./ui/input";
+import { Label } from "./ui/label";
+import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "./ui/select";
+import { Skeleton } from "./ui/skeleton";
+import { Mieter } from "../lib/data-fetching";
+import { BerechnungsartValue, BERECHNUNGSART_OPTIONS } from "../lib/constants";
+
+export interface CostItem {
+  id: string;
+  art: string;
+  betrag: string;
+  berechnungsart: BerechnungsartValue | '';
+}
+
+export interface RechnungEinzel {
+  mieterId: string;
+  betrag: string;
+}
+
+export interface SortableCostItemProps {
+  item: CostItem;
+  index: number;
+  costItems: CostItem[];
+  selectedHausMieter: Mieter[];
+  rechnungen: Record<string, RechnungEinzel[]>;
+  isSaving: boolean;
+  isLoadingDetails: boolean;
+  isFetchingTenants: boolean;
+  hausId: string;
+  onCostItemChange: (index: number, field: keyof Omit<CostItem, 'id'>, value: string | BerechnungsartValue) => void;
+  onRemoveCostItem: (index: number) => void;
+  onRechnungChange: (costItemId: string, mieterId: string, newBetrag: string) => void;
+  hoveredBerechnungsart: BerechnungsartValue | '';
+  selectContentRect: DOMRect | null;
+  hoveredItemRect: DOMRect | null;
+  tooltipMap: Record<BerechnungsartValue | '', string>;
+  onItemHover: (e: React.MouseEvent<HTMLElement> | React.FocusEvent<HTMLElement>, value: BerechnungsartValue) => void;
+  onItemLeave: () => void;
+  selectContentRef: React.RefObject<HTMLDivElement | null>;
+}
+
+const normalizeBerechnungsart = (rawValue: string): BerechnungsartValue => {
+  const berechnungsartMap: Record<string, BerechnungsartValue> = {
+    'pro person': 'pro Mieter',
+    'pro mieter': 'pro Mieter',
+    'pro flaeche': 'pro Flaeche',
+    'pro fläche': 'pro Flaeche',
+    'pro qm': 'pro Flaeche',
+    'qm': 'pro Flaeche',
+    'pro wohnung': 'pro Wohnung',
+    'nach rechnung': 'nach Rechnung',
+  };
+  
+  const lower = rawValue.toLowerCase();
+  const normalized = berechnungsartMap[lower] || rawValue;
+  return (BERECHNUNGSART_OPTIONS.find(opt => opt.value === normalized)?.value as BerechnungsartValue) || '';
+};
+
+export function SortableCostItem({
+  item,
+  index,
+  costItems,
+  selectedHausMieter,
+  rechnungen,
+  isSaving,
+  isLoadingDetails,
+  isFetchingTenants,
+  hausId,
+  onCostItemChange,
+  onRemoveCostItem,
+  onRechnungChange,
+  hoveredBerechnungsart,
+  selectContentRect,
+  hoveredItemRect,
+  tooltipMap,
+  onItemHover,
+  onItemLeave,
+  selectContentRef,
+}: SortableCostItemProps) {
+  const {
+    attributes,
+    listeners,
+    setNodeRef,
+    transform,
+    transition,
+    isDragging,
+  } = useSortable({ id: item.id });
+
+  const style = {
+    transform: CSS.Transform.toString(transform),
+    transition,
+    opacity: isDragging ? 0.5 : 1,
+  };
+
+  return (
+    <div
+      ref={setNodeRef}
+      style={style}
+      className={`flex flex-col gap-3 py-2 border-b last:border-b-0 ${isDragging ? 'z-10' : ''}`}
+      role="group"
+      aria-label={`Kostenposition ${index + 1}`}
+    >
+      <div className="flex flex-col sm:flex-row items-start gap-3">
+        <div className="flex items-center justify-center flex-none w-8 h-10">
+          <button
+            type="button"
+            className="cursor-grab active:cursor-grabbing p-1 hover:bg-gray-100 rounded transition-colors flex items-center justify-center w-6 h-6"
+            {...attributes}
+            {...listeners}
+            aria-label="Kostenposition verschieben"
+          >
+            <GripVertical className="h-4 w-4 text-gray-400" />
+          </button>
+        </div>
+        <div className="w-full sm:flex-[4_1_0%]">
+          <Input
+            id={`art-${item.id}`}
+            placeholder="Kostenart"
+            value={item.art}
+            onChange={(e) => onCostItemChange(index, 'art', e.target.value)}
+            disabled={isSaving}
+          />
+        </div>
+        <div className="w-full sm:flex-[3_1_0%]">
+          {item.berechnungsart === 'nach Rechnung' ? (
+            <div className="flex items-center justify-center h-10 px-3 py-2 text-sm text-muted-foreground bg-gray-50 border rounded-md">
+              Beträge pro Mieter unten
+            </div>
+          ) : (
+            <Input
+              id={`betrag-${item.id}`}
+              type="number"
+              placeholder="Betrag (€)"
+              value={item.betrag}
+              onChange={(e) => onCostItemChange(index, 'betrag', e.target.value)}
+              step="0.01"
+              disabled={isSaving}
+            />
+          )}
+        </div>
+        <div className="w-full sm:flex-[4_1_0%]">
+          <Select
+            value={item.berechnungsart}
+            onValueChange={(value) => onCostItemChange(index, 'berechnungsart', value as BerechnungsartValue)}
+            disabled={isSaving}
+          >
+            <SelectTrigger
+              onMouseEnter={(e) => item.berechnungsart && onItemHover(e, item.berechnungsart as BerechnungsartValue)}
+              onMouseLeave={onItemLeave}
+              onFocus={(e) => item.berechnungsart && onItemHover(e, item.berechnungsart as BerechnungsartValue)}
+              onBlur={onItemLeave}
+              ref={selectContentRef as any}
+            >
+              <SelectValue placeholder="Berechnungsart" />
+            </SelectTrigger>
+            <SelectContent>
+              {BERECHNUNGSART_OPTIONS.map((option) => (
+                <SelectItem
+                  key={option.value}
+                  value={option.value}
+                  onMouseEnter={(e: React.MouseEvent<HTMLDivElement>) => onItemHover(e as unknown as React.MouseEvent<HTMLDivElement>, option.value as BerechnungsartValue)}
+                  onMouseLeave={onItemLeave}
+                >
+                  {option.label}
+                </SelectItem>
+              ))}
+            </SelectContent>
+          </Select>
+          {selectContentRect && hoveredBerechnungsart && hoveredItemRect && createPortal(
+            <div
+              className="fixed z-[60] transition-none"
+              style={{
+                top: `${Math.round(hoveredItemRect.top)}px`,
+                right: `${window.innerWidth - Math.round(selectContentRect.left) + 8}px`,
+                width: '280px',
+                minHeight: `${Math.round(hoveredItemRect.height)}px`,
+              }}
+            >
+              <div className="h-full rounded-md border bg-popover text-popover-foreground shadow-sm p-3 text-sm flex items-center">
+                {tooltipMap[hoveredBerechnungsart]}
+              </div>
+            </div>,
+            document.body
+          )}
+        </div>
+        <div className="flex items-center justify-center flex-none w-10 h-10">
+          <Button
+            type="button"
+            variant="ghost"
+            size="icon"
+            onClick={() => onRemoveCostItem(index)}
+            disabled={costItems.length <= 1 || isLoadingDetails || isSaving}
+            aria-label="Kostenposition entfernen"
+            className="h-8 w-8"
+          >
+            <Trash2 className="h-4 w-4 text-destructive" />
+          </Button>
+        </div>
+      </div>
+
+      {item.berechnungsart === 'nach Rechnung' && (
+        <div className="mt-3 p-4 bg-gray-50 border border-gray-200 rounded-md space-y-3 shadow-sm">
+          <h4 className="text-md font-semibold text-gray-700">
+            Einzelbeträge für: <span className="font-normal italic">"{item.art || 'Unbenannte Kostenart'}"</span>
+          </h4>
+          {isFetchingTenants ? (
+            Array.from({ length: 3 }).map((_, skelIdx) => (
+              <div key={`skel-tenant-${skelIdx}`} className="grid grid-cols-10 gap-2 items-center py-1">
+                <Skeleton className="h-8 w-full col-span-6 sm:col-span-7" />
+                <Skeleton className="h-8 w-full col-span-4 sm:col-span-3" />
+              </div>
+            ))
+          ) : (
+            <>
+              {!isFetchingTenants && !hausId && (
+                <p className="text-sm text-orange-600 p-2 bg-orange-50 border border-orange-200 rounded-md">
+                  Bitte wählen Sie zuerst ein Haus aus, um Mieter zu laden.
+                </p>
+              )}
+              {!isFetchingTenants && hausId && selectedHausMieter.length === 0 && !isLoadingDetails && (
+                <p className="text-sm text-orange-600 p-2 bg-orange-50 border border-orange-200 rounded-md">
+                  Für das ausgewählte Haus wurden keine Mieter gefunden.
+                </p>
+              )}
+              {selectedHausMieter.length > 0 && (
+                <div className="space-y-2 max-h-60 overflow-y-auto pr-2">
+                  {selectedHausMieter.map(mieter => {
+                    const rechnungForMieter = (rechnungen[item.id] || []).find(r => r.mieterId === mieter.id);
+                    return (
+                      <div key={mieter.id} className="grid grid-cols-10 gap-2 items-center py-1 border-b border-gray-100 last:border-b-0">
+                        <Label htmlFor={`rechnung-${item.id}-${mieter.id}`} className="col-span-6 sm:col-span-7 truncate text-sm" title={mieter.name}>
+                          {mieter.name}
+                        </Label>
+                        <div className="col-span-4 sm:col-span-3">
+                          <Input
+                            id={`rechnung-${item.id}-${mieter.id}`}
+                            type="number"
+                            step="0.01"
+                            placeholder="Betrag (€)"
+                            value={rechnungForMieter?.betrag || ''}
+                            onChange={(e) => onRechnungChange(item.id, mieter.id, e.target.value)}
+                            className="w-full text-sm"
+                            disabled={isLoadingDetails || isSaving}
+                          />
+                        </div>
+                      </div>
+                    );
+                  })}
+                </div>
+              )}
+              {rechnungen[item.id] && selectedHausMieter.length > 0 && (
+                <div className="pt-2 mt-2 border-t border-gray-300 flex justify-end">
+                  <p className="text-sm font-semibold text-gray-700">
+                    Summe: {formatNumber((rechnungen[item.id] || []).reduce((sum, r) => sum + (parseFloat(r.betrag) || 0), 0))} €
+                  </p>
+                </div>
+              )}
+            </>
+          )}
+        </div>
+      )}
+    </div>
+  );
+}
